Fixed the problem I was having with trailer thumbnails not all loading in the detail page by changing the list of trailer thumbs to be a RecyclerView rather than ListView
I believe that ListView was problematic for the picasso thumbs because ListView had problems loading any image off the screen. Maybe there was a workaround to use ListView with Picasso thumbs, but I was able to get it up and running with a RecyclerView instead very easily. I also added onClickListeners to the trailer thumbnails to launch the youtube app in order to play the trailer.