Stuck.
I'm getting a NPE because I'm trying to set the supportactionbar from the MovieDetailActivity. This is because now I'm setting the content view to `movie_detail_activity`, which is just a frame layout, instead of `movie_detail_fragment`, which actually contains the toolbar.

Wondering if there's a way to add a fragment without needing the frame in the detail OR if I can somehow access the toolbar from the activity